Move Dask crash reconciliation into core futures
