+++ src/integrations/prefect-dask/tests/test_task_runners.py
@@ -1,10 +1,8 @@
 import asyncio
 import time
-from types import SimpleNamespace
 from typing import Generator, List
-from unittest.mock import MagicMock
-from uuid import uuid4
 
+import dask
 import dask.dataframe as dd
 import distributed
 import pandas as pd
@@ -16,7 +14,6 @@ from prefect_dask.task_runners import PrefectDaskFuture
 from prefect import flow, task
 from prefect.assets import Asset, materialize
 from prefect.client.orchestration import get_client
-from prefect.client.schemas.responses import SetStateStatus
 from prefect.context import get_run_context
 from prefect.futures import as_completed
 from prefect.server.schemas.states import StateType
@@ -64,63 +61,6 @@ def default_dask_task_runner():  # noqa
             "dashboard_address": None,  # Prevent port conflicts
         }
     )
-
-
-class TestPrefectDaskFuture:
-    def test_wait_marks_scheduler_exception_as_crashed(self, monkeypatch):
-        task_run_id = uuid4()
-        wrapped_future = MagicMock()
-        wrapped_future.result.side_effect = RuntimeError("worker exited unexpectedly")
-
-        proposed_state = State(
-            type=StateType.CRASHED,
-            message="Execution was interrupted by an unexpected exception: RuntimeError: worker exited unexpectedly",
-        )
-        client = MagicMock()
-        client.set_task_run_state.return_value = SimpleNamespace(
-            status=SetStateStatus.ACCEPT,
-            state=proposed_state,
-            details=None,
-        )
-
-        monkeypatch.setattr("prefect_dask.task_runners.get_client", lambda **_: client)
-
-        future = PrefectDaskFuture(
-            wrapped_future=wrapped_future, task_run_id=task_run_id
-        )
-
-        future.wait()
-
-        client.set_task_run_state.assert_called_once()
-        assert future.state.type == StateType.CRASHED
-
-    def test_result_marks_scheduler_exception_as_crashed(self, monkeypatch):
-        task_run_id = uuid4()
-        wrapped_future = MagicMock()
-        wrapped_future.result.side_effect = RuntimeError("worker exited unexpectedly")
-
-        proposed_state = State(
-            type=StateType.CRASHED,
-            message="Execution was interrupted by an unexpected exception: RuntimeError: worker exited unexpectedly",
-        )
-        client = MagicMock()
-        client.set_task_run_state.return_value = SimpleNamespace(
-            status=SetStateStatus.ACCEPT,
-            state=proposed_state,
-            details=None,
-        )
-
-        monkeypatch.setattr("prefect_dask.task_runners.get_client", lambda **_: client)
-
-        future = PrefectDaskFuture(
-            wrapped_future=wrapped_future, task_run_id=task_run_id
-        )
-
-        with pytest.raises(Exception):
-            future.result()
-
-        client.set_task_run_state.assert_called_once()
-        assert future.state.type == StateType.CRASHED
 
 
 class TestDaskTaskRunner:
@@ -361,6 +301,31 @@ class TestDaskTaskRunner:
                 wait_for=[],
             )
 
+            future.wait()
+            assert future.state.type == StateType.CRASHED
+
+    def test_worker_loss_exhaustion_marks_task_run_crashed(self):
+        dask.config.set({"distributed.scheduler.allowed-failures": 1})
+
+        @task
+        def oom_task():
+            blocks = []
+            while True:
+                blocks.append(bytearray(64 * 1024 * 1024))
+                time.sleep(0.05)
+
+        task_runner = DaskTaskRunner(
+            cluster_kwargs={
+                "processes": True,
+                "n_workers": 1,
+                "threads_per_worker": 1,
+                "dashboard_address": None,
+                "memory_limit": "256 MiB",
+            }
+        )
+
+        with task_runner:
+            future = task_runner.submit(oom_task, parameters={}, wait_for=[])
             future.wait()
             assert future.state.type == StateType.CRASHED
